Move fetch and redux action to app
